Refactor slugy_test.ex
* Moved support modules to test/support folder
