+++ test/slugy_test.exs
@@ -2,41 +2,7 @@ defmodule SlugyTest do
   use ExUnit.Case
 
   alias Ecto.Changeset
-  alias Slugy.Support.Content
-
-  defmodule Post do
-    use Ecto.Schema
-    import Ecto.Changeset
-    import Slugy, only: [slugify: 2]
-
-    embedded_schema do
-      field(:title, :string)
-      field(:slug, :string)
-    end
-
-    def changeset(post, attrs) do
-      post
-      |> cast(attrs, [:title])
-      |> slugify(:title)
-    end
-  end
-
-  defmodule PostWithEmbeddedStruct do
-    use Ecto.Schema
-    import Ecto.Changeset
-    import Slugy, only: [slugify: 2]
-
-    embedded_schema do
-      field(:data, :map) # => %PostWithEmbeddedStruct{data: %{title: "A title"}}
-      field(:slug, :string)
-    end
-
-    def changeset(post, attrs) do
-      post
-      |> cast(attrs, [:data])
-      |> slugify([:data, :title])
-    end
-  end
+  alias Slugy.Support.{Content, Post, PostWithEmbeddedStruct}
 
   doctest Slugy
